장고수업/1017/08_django/02_drf/requirements.txt
asgiref==3.5.2
Django==3.2.13
django-extensions==3.2.1
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3
